Add temporary prime check to get tests passing
Adopted the explicit Given, When, Then syntax in tests to clarify
intention.

I also wanted to add a naive implementation for finding primes here to
make sure I understand the full problem and how the multiplication
tables are going to be generated before I dive into the details of an
algorithm to find primes.